Do not reacquire project lock when completing a worker-based task
The project lock is released if submitted work items are pending when the
task action completes. With this change, the lock is no longer reacquired
once these work items finish: instead, the task is marked as "complete"
without reacquiring the project-mutation lock.

This isn't quite correct, since there are cases where the project lock
should be reacquired, such as when a task as further task actions to execute.
This will be addressed in a subsequent commit.